Create /blog/feed.xml with feature parity of old atom.xml
Decided that /blog/feed.xml makes more sense than having it at the root
as the /blog is a section of the website for blog posts. Root level
feeds may imply something else in the future.

In order for images to be viewable in feed readers, the imgs linked to
must be full URLs.

To do this, we copied the RSS layout from this helpful blog post:
https://retifrav.github.io/blog/2019/03/17/migrating-from-octopress-to-hugo/#rss

and hardcoded the number of posts to list to 20.

Signed-off-by: Kool User <lx4r@users.noreply.github.com>